update the main.js and style.css
