+++ style.css
@@ -1,0 +1,46 @@
+* {
+    margin: 0;
+    padding: 0;
+  }
+  
+  body {
+    text-align: center;
+  }
+  
+  .book {
+    display: flex;
+    justify-content: space-between;
+    max-width: 600px;
+    margin: 0 auto;
+    padding: 20px;
+  }
+  
+  .dark {
+    background-color: rgba(51, 51, 51, 0.48);
+  }
+  
+  button {
+    padding: 8px;
+    cursor: pointer;
+  }
+  
+  form {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    width: 400px;
+    margin: 0 auto;
+    padding: 20px;
+  }
+  
+  form input{
+    padding: 8px;
+    margin: 8px;
+    width: 100%;
+    border: 1px solid #ccc;
+  }
+  
+  form button {
+    align-self: flex-end;
+  }
